CodeQL's first scan is answered, in code where it can be
ADR-0341. The first scheduled security-and-quality run found 184 things
in tracked files; reproduced locally with the same CLI and suite, since
the alerts API needs a token, and every one of them is now either fixed
or on the record with its reason. No rule and no path is excluded.

Fixed: the four security findings, int loop counters run against long
and double bounds, in the golden resampler, the clipboard MIME loop and
TimeTicks, which now names the label count it compares. Path's verb
stride is a switch rather than a byte-indexed array; Transform.mix takes
each side from its own list or the other's identity with no null path;
ParagraphCache's eviction hook says super.size(). Where a number came
from somebody -- the launcher's flags, an icon's point list, a
document's action argument, the showcase's task toggle -- a
NumberFormatException becomes a refusal that names what was asked.

The 61 bindings spelled `ignored` become JDK 22's `_`, and seven lambda
parameters with them; palantir 2.97 needs `var _` inside a record
pattern, so that is the spelling there. Three dead parameters go. Two
tests that asserted nothing -- ImmutabilityTest's handlers, and the
select popup's marked row -- assert what their names claim.

What stays: the inflate factories' contract parameters, the upcall
targets' C signatures, nine compact constructors that already copy,
three multi-label switch arms, four validated CSS parses, and the `_`
bindings themselves -- CodeQL does not read JEP 456 yet, on 2.27 or on
main. docs/testing.md §2 carries the tokenless reproduction recipe.

+++ natives/src/main/java/io/github/digitalsmile/goldberry/natives/sdl/desktop/SdlClipboard.java
@@ -411,8 +411,10 @@ public final class SdlClipboard {
             return List.of();
         }
         var pointers = array.reinterpret(count * ValueLayout.ADDRESS.byteSize());
-        var types = new java.util.ArrayList<String>((int) count);
-        for (var i = 0; i < count; i++) {
+        var types = new java.util.ArrayList<String>(Math.toIntExact(count));
+        // A long counter for a long count. SDL will never report two billion
+        // MIME types, but an int here would wrap rather than stop if it did.
+        for (var i = 0L; i < count; i++) {
             var pointer = pointers.getAtIndex(ValueLayout.ADDRESS, i);
             if (!MemorySegment.NULL.equals(pointer)) {
                 types.add(readCString(pointer));
